testing.

+++ ion/agents/instrument/driver_process.py
@@ -136,15 +136,25 @@ class DriverProcess(object):
 
             if not force and self._driver_client:
                 try:
-                    log.info('Stopping driver process.')
+                    log.info('Stopping driver process.') 
                     self._driver_client.done()
                     self._driver_process.wait()
                     log.info('Driver process stopped.')
                 except:
                     try:
                         log.error('Exception stopping driver process...killing')
-                        self._driver_process.kill()
-                        self._driver_process.wait()
+                        print '## polling'
+                        self._driver_process.poll()
+                        if not self._driver_process.returncode:
+                            print '## poll return code is: ' + str(self._driver_process.returncode)
+                            print '## killing'
+                            self._driver_process.kill()
+                            self._driver_process.wait()
+                        print '## stopping'
+                        #self._driver_client.stop_messaging()
+                        #self._driver_client.stop_event_thread = True
+                        #self._driver_client.event_thread.kill()
+                        #self._driver_client.event_thread.join()
                         log.error('Driver process killed.')
                     except Exception as ex:
                         log.error('Exception killing driver process')
@@ -154,8 +164,18 @@ class DriverProcess(object):
             else:
                 try:
                     log.info('Killing driver process.')
-                    self._driver_process.kill()
-                    self._driver_process.wait()
+                    print '## polling'
+                    self._driver_process.poll()
+                    if  not self._driver_process.returncode:
+                        print '## poll return code is: ' + str(self._driver_process.returncode)
+                        print '## killing'
+                        self._driver_process.kill()
+                        self._driver_process.wait()
+                    print '## stopping'
+                    self._driver_client.stop_messaging()
+                    #self._driver_client.stop_event_thread = True
+                    #self._driver_client.event_thread.kill()
+                    #self._driver_client.event_thread.join()
                     log.info('Driver process killed.')
                 except Exception as ex:
                     log.error('Exception killing driver process')
